+++ servo/connectors/__init__.py
@@ -1,0 +1,1 @@
+"""Contains connectors that are bundled with the servo core package."""